Split linux-only config code
