+++ pkg/gpu/config/config.go
@@ -8,13 +8,11 @@ package config
 
 import (
 	"errors"
-	"fmt"
 	"time"
 
 	sysconfig "github.com/DataDog/datadog-agent/cmd/system-probe/config"
 	pkgconfigsetup "github.com/DataDog/datadog-agent/pkg/config/setup"
 	"github.com/DataDog/datadog-agent/pkg/ebpf"
-	"github.com/DataDog/datadog-agent/pkg/util/kernel"
 )
 
 // GPUNS is the namespace for the GPU monitoring probe.
@@ -22,28 +20,6 @@ const GPUNS = "gpu_monitoring"
 
 // ErrNotSupported is the error returned if GPU monitoring is not supported on this platform
 var ErrNotSupported = errors.New("GPU Monitoring is not supported")
-
-// MinimumKernelVersion indicates the minimum kernel version required for GPU monitoring
-var MinimumKernelVersion kernel.Version
-
-func init() {
-	// we rely on ring buffer support for GPU monitoring, hence the minimal kernel version is 5.8.0
-	MinimumKernelVersion = kernel.VersionCode(5, 8, 0)
-}
-
-// CheckGPUSupported checks if the host's kernel supports GPU monitoring
-func CheckGPUSupported() error {
-	kversion, err := kernel.HostVersion()
-	if err != nil {
-		return fmt.Errorf("%w: could not determine the current kernel version: %w", ErrNotSupported, err)
-	}
-
-	if kversion < MinimumKernelVersion {
-		return fmt.Errorf("%w: a Linux kernel version of %s or higher is required; we detected %s", ErrNotSupported, MinimumKernelVersion, kversion)
-	}
-
-	return nil
-}
 
 // Config holds the configuration for the GPU monitoring probe.
 type Config struct {
